Allow use of unsigned requests to S3
For public S3 buckets, credentials are not required. This commit adds a
new configuration parameter: 'aws_s3_no_sign_request' which will make
the S3 requests without using credentials if set to 'true'. It is
equivalent to the AWS CLI option --no-sign-request.

The boto library first checks for this option before looking for
credentials meaning if this option is set to true, it will override any
credentials that have been set.