Disallow unknown commandline arguments
Previously commandline arguments that were provided but not 'known' to
HSDS were ignored. This can create confusing situations where, for
example the --config_dir argument is passed as --config-dir (note
underscore vs hyphen) which then leads to trying to understand
which part of the config is incorrect.

Instead report that an unknown argument has been provided.