Use K8s API label selector to filter pods
Previously, the filter logic was hardcoded as looking up for a
single pod label with a fixed
name (`app`) that matched the value set in the `k8s_app_label` config.

Now, a new `k8s_dn_label_selector` config is provided to allow the user
to provide a K8s label selector. This selector is then added as a
parameter in the K8s API request.

The use of a label selector allows the user more flexibility to choose the
label name, and even to filter by more than one label.

For backward compatibility, the old `k8s_app_label` config is still
supported if no `k8s_dn_label_selector` is given.

Signed-off-by: Thiago M. de C. Marques <thiago_manhente@yahoo.com.br>