Skip Kubernetes tests by default
Previously, the test `setUp()` tried to call `kubectl create namespace` and
skipped the test if no `kubectl` command was found.

The GitHub Workflow environment, however, has the `kubectl` command but no
cluster connected to use. The `kubectl` call made by the test fails with exit
code 1 and a message _The connection to the server localhost:8080 was refused"_.
The tests then fail with a `CalledProcessError` exception.

To avoid that, the `K8sListDnTest` is now skipped by default. The test must be
ran manually on an environment that has a K8s cluster suitable for creating the
test pods and services.